+++ swagger-request-validator-core/src/main/java/com/atlassian/oai/validator/model/SimpleRequest.java
@@ -5,7 +5,10 @@ import com.google.common.collect.MultimapBuilder;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
-import java.util.Arrays;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+import java.time.format.DateTimeParseException;
+import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
@@ -41,7 +44,7 @@ public class SimpleRequest implements Request {
         this.path = requireNonNull(path, "A request path is required");
         this.queryParams = requireNonNull(queryParams);
         this.headers = requireNonNull(headers);
-        this.requestBody = Optional.ofNullable(body);
+        requestBody = Optional.ofNullable(body);
     }
 
     @Nonnull
@@ -239,8 +242,8 @@ public class SimpleRequest implements Request {
             this.method = requireNonNull(method, "A method is required");
             this.path = requireNonNull(path, "A path is required");
 
-            this.headers = multimapBuilder(false /* header are always case insensitive */);
-            this.queryParams = multimapBuilder(queryParametersCaseSensitive);
+            headers = multimapBuilder(false /* header are always case insensitive */);
+            queryParams = multimapBuilder(queryParametersCaseSensitive);
         }
 
         /**
@@ -341,7 +344,34 @@ public class SimpleRequest implements Request {
         }
 
         static Collection<String> splitHeaderValue(final String value) {
-            return value != null ? Arrays.asList(value.split("\\s*,\\s*")) : Collections.singleton(null);
+            if (value == null) {
+                return Collections.singleton(null);
+            }
+            final String[] split = value.split("\\s*,\\s*");
+
+            //check if dates in RFC_1123
+            final List<String> result = new ArrayList<>();
+            int i = 0;
+            while (i < split.length) {
+                // Last element in the split
+                if (i == split.length - 1) {
+                    result.add(split[i]);
+                    break;
+                }
+
+                try {
+                    // Check to see if we can make an RFC_1123 Date from the next two items
+                    final String possibleDate = split[i] + ", " + split[i + 1];
+                    LocalDateTime.parse(split[i] + ", " + split[i + 1], DateTimeFormatter.RFC_1123_DATE_TIME);
+                    result.add(possibleDate);
+                    i += 2;
+                } catch (final DateTimeParseException e) {
+                    // If not, add the item and continue
+                    result.add(split[i]);
+                    i += 1;
+                }
+            }
+            return result;
         }
     }
 }