+++ swagger-request-validator-springmvc/src/main/java/com/atlassian/oai/validator/springmvc/SwaggerValidationInterceptor.java
@@ -1,5 +1,6 @@
 package com.atlassian.oai.validator.springmvc;
 
+import com.atlassian.oai.validator.SwaggerRequestResponseValidator;
 import com.atlassian.oai.validator.model.Request;
 import com.atlassian.oai.validator.model.Response;
 import com.atlassian.oai.validator.report.ValidationReport;
@@ -25,6 +26,10 @@ public class SwaggerValidationInterceptor extends HandlerInterceptorAdapter {
 
     public SwaggerValidationInterceptor(final EncodedResource swaggerInterface) throws IOException {
         this(new SwaggerRequestValidationService(swaggerInterface));
+    }
+
+    public SwaggerValidationInterceptor(final SwaggerRequestResponseValidator swaggerRequestResponseValidator) {
+        this(new SwaggerRequestValidationService(swaggerRequestResponseValidator));
     }
 
     SwaggerValidationInterceptor(final SwaggerRequestValidationService swaggerRequestValidationService) {
@@ -62,9 +67,6 @@ public class SwaggerValidationInterceptor extends HandlerInterceptorAdapter {
         final ValidationReport validationReport = swaggerRequestValidationService.validateRequest(request);
         if (!validationReport.hasErrors()) {
             LOG.debug("Swagger request validation: {} - The request is valid.", requestLoggingKey);
-        } else if (!swaggerRequestValidationService.isDefinedSwaggerRequest(validationReport)) {
-            LOG.info("Swagger request validation: {} - The request is not defined in the Swagger schema. Ignoring it.",
-                    requestLoggingKey);
         } else {
             final InvalidRequestException invalidRequestException = new InvalidRequestException(validationReport);
             LOG.info("Swagger request validation: {} - The request is invalid: {}", requestLoggingKey,
@@ -109,9 +111,6 @@ public class SwaggerValidationInterceptor extends HandlerInterceptorAdapter {
         final ValidationReport validationReport = swaggerRequestValidationService.validateResponse(servletRequest, response);
         if (!validationReport.hasErrors()) {
             LOG.debug("Swagger response validation: {} - The response is valid.", requestLoggingKey);
-        } else if (!swaggerRequestValidationService.isDefinedSwaggerRequest(validationReport)) {
-            LOG.info("Swagger response validation: {} - The request is not defined in the Swagger schema. Ignoring it.",
-                    requestLoggingKey);
         } else {
             final InvalidResponseException invalidResponseException = new InvalidResponseException(validationReport);
             LOG.info("Swagger response validation: {} - The response is invalid: {}", requestLoggingKey,