+++ swagger-request-validator-springmvc/src/test/java/com/atlassian/oai/validator/springmvc/SwaggerValidationInterceptorTest.java
@@ -1,5 +1,6 @@
 package com.atlassian.oai.validator.springmvc;
 
+import com.atlassian.oai.validator.SwaggerRequestResponseValidator;
 import com.atlassian.oai.validator.model.Request;
 import com.atlassian.oai.validator.model.Response;
 import com.atlassian.oai.validator.report.ValidationReport;
@@ -18,6 +19,7 @@ import java.util.Collections;
 
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.notNullValue;
+import static org.junit.Assert.assertThat;
 import static org.mockito.Mockito.times;
 
 public class SwaggerValidationInterceptorTest {
@@ -39,6 +41,19 @@ public class SwaggerValidationInterceptorTest {
 
         final SwaggerValidationInterceptor interceptor = new SwaggerValidationInterceptor(encodedResource);
         Assert.assertThat(interceptor, notNullValue());
+    }
+
+    @Test
+    public void constructor_withSwaggerRequestResponseValidator() {
+        // given:
+        final SwaggerRequestResponseValidator swaggerRequestResponseValidator =
+                Mockito.mock(SwaggerRequestResponseValidator.class);
+
+        // when:
+        final SwaggerValidationInterceptor interceptor = new SwaggerValidationInterceptor(swaggerRequestResponseValidator);
+
+        // then:
+        assertThat(interceptor, notNullValue());
     }
 
     @Test
@@ -68,26 +83,6 @@ public class SwaggerValidationInterceptorTest {
         Assert.assertThat(result, equalTo(true));
     }
 
-    @Test
-    public void preHandle_theRequestIsNotPartOfTheSwaggerDefinition() throws Exception {
-        final HttpServletRequest servletRequest = Mockito.mock(ResettableRequestServletWrapper.class);
-        final Request request = Mockito.mock(Request.class);
-        final ValidationReport validationReport = Mockito.mock(ValidationReport.class);
-
-        Mockito.when(servletRequest.getMethod()).thenReturn("METHOD");
-        Mockito.when(servletRequest.getRequestURI()).thenReturn("/request/uri");
-
-        Mockito.when(swaggerRequestValidationService.buildRequest(servletRequest)).thenReturn(request);
-        Mockito.when(swaggerRequestValidationService.validateRequest(request)).thenReturn(validationReport);
-        Mockito.when(validationReport.hasErrors()).thenReturn(true);
-        Mockito.when(swaggerRequestValidationService.isDefinedSwaggerRequest(validationReport)).thenReturn(false);
-
-        final boolean result = classUnderTest.preHandle(servletRequest, null, null);
-
-        Mockito.verify(swaggerRequestValidationService, times(1)).isDefinedSwaggerRequest(validationReport);
-        Assert.assertThat(result, equalTo(true));
-    }
-
     @Test(expected = InvalidRequestException.class)
     public void preHandle_theRequestIsInvalid() throws Exception {
         final HttpServletRequest servletRequest = Mockito.mock(ResettableRequestServletWrapper.class);
@@ -100,7 +95,6 @@ public class SwaggerValidationInterceptorTest {
         Mockito.when(swaggerRequestValidationService.buildRequest(servletRequest)).thenReturn(request);
         Mockito.when(swaggerRequestValidationService.validateRequest(request)).thenReturn(validationReport);
         Mockito.when(validationReport.hasErrors()).thenReturn(true);
-        Mockito.when(swaggerRequestValidationService.isDefinedSwaggerRequest(validationReport)).thenReturn(true);
         Mockito.when(validationReport.getMessages()).thenReturn(Collections.emptyList());
 
         final boolean result = classUnderTest.preHandle(servletRequest, null, null);
@@ -141,30 +135,6 @@ public class SwaggerValidationInterceptorTest {
         Mockito.verify(validationReport, times(1)).hasErrors();
     }
 
-    @Test
-    public void postHandle_theRequestIsNotPartOfTheSwaggerDefinition() throws Exception {
-        // given:
-        final HttpServletRequest servletRequest = Mockito.mock(HttpServletRequest.class);
-        final ContentCachingResponseWrapper servletResponse = Mockito.mock(ContentCachingResponseWrapper.class);
-        final Response response = Mockito.mock(Response.class);
-        final ValidationReport validationReport = Mockito.mock(ValidationReport.class);
-
-        // and:
-        Mockito.when(servletRequest.getMethod()).thenReturn("METHOD");
-        Mockito.when(servletRequest.getRequestURI()).thenReturn("/request/uri");
-
-        Mockito.when(swaggerRequestValidationService.buildResponse(servletResponse)).thenReturn(response);
-        Mockito.when(swaggerRequestValidationService.validateResponse(servletRequest, response)).thenReturn(validationReport);
-        Mockito.when(validationReport.hasErrors()).thenReturn(true);
-        Mockito.when(swaggerRequestValidationService.isDefinedSwaggerRequest(validationReport)).thenReturn(false);
-
-        // when:
-        classUnderTest.postHandle(servletRequest, servletResponse, null, null);
-
-        // then:
-        Mockito.verify(swaggerRequestValidationService, times(1)).isDefinedSwaggerRequest(validationReport);
-    }
-
     @Test(expected = InvalidResponseException.class)
     public void postHandle_theResponseIsInvalid() throws Exception {
         // setup:
@@ -180,7 +150,6 @@ public class SwaggerValidationInterceptorTest {
         Mockito.when(swaggerRequestValidationService.buildResponse(servletResponse)).thenReturn(response);
         Mockito.when(swaggerRequestValidationService.validateResponse(servletRequest, response)).thenReturn(validationReport);
         Mockito.when(validationReport.hasErrors()).thenReturn(true);
-        Mockito.when(swaggerRequestValidationService.isDefinedSwaggerRequest(validationReport)).thenReturn(true);
         Mockito.when(validationReport.getMessages()).thenReturn(Collections.emptyList());
 
         // when: